Move ProjectViewSet from api.py to views.py
