+++ mysite/projects/urls.py
@@ -1,5 +1,5 @@
 from rest_framework import routers
-from .api import ProjectViewSet
+from .views import ProjectViewSet
 
 from django.urls import path, include
 from . import views
