+++ mysite/projects/views.py
@@ -1,10 +1,15 @@
-from django.shortcuts import render
-from rest_framework import generics
+from rest_framework import generics, viewsets, permissions
 
 from .models import Project
 from .serializers import ProjectSerializer
 # Create your views here.
 
+class ProjectViewSet(viewsets.ModelViewSet):
+    queryset = Project.objects.all()
+    permission_classes = [
+        permissions.AllowAny
+    ]
+    serializer_class = ProjectSerializer
 
 class ProjectListCreate(generics.ListCreateAPIView):
     queryset = Project.objects.all()
